pyproject.toml
[project]
name = "federated-deployment"
version = "0.1.0"
description = "FedCost: Federated Learning for DRG-Based Healthcare Cost Prediction — AWS Infrastructure"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.34",
    "matplotlib>=3.10.8",
    "numpy>=1.26",
    "pandas>=2.2",
    "pulumi>=3.0",
    "pulumi-aws>=6.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0",
    "scikit-learn>=1.4",
    "torch>=2.2",
    "xgboost>=3.2.0",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "boto3>=1.34",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
